Add exception class for CheckedIssueAnalyzer and fix bugs for open files
