+++ config.properties
@@ -1,0 +1,4 @@
+url=https://github.com/google/guava
+path=/home/mialyshev/guava
+projectname=guava
+issuespath=scan.txt
