+++ pom.xml
@@ -46,6 +46,11 @@
       <artifactId>json</artifactId>
       <version>20190722</version>
     </dependency>
+    <dependency>
+      <groupId>org.slf4j</groupId>
+      <artifactId>slf4j-nop</artifactId>
+      <version>1.7.13</version>
+    </dependency>
   </dependencies>
 
   <build>
@@ -69,6 +74,7 @@
                     <exclude>META-INF/*.SF</exclude>
                     <exclude>META-INF/*.DSA</exclude>
                     <exclude>META-INF/*.RSA</exclude>
+                    <exclude>**/*.properties</exclude>
                   </excludes>
                 </filter>
               </filters>
