+++ src/main/java/org/suai/blamer/Main.java
@@ -1,6 +1,8 @@
 package org.suai.blamer;
 
 import org.apache.commons.cli.*;
+import org.suai.blamer.check.CheckedIssueAnalyzer;
+import org.suai.blamer.check.CheckedIssueException;
 import org.suai.blamer.git.BlameInspector;
 import org.suai.blamer.git.GitException;
 import org.suai.blamer.issuetracker.GithubIssueManager;
@@ -8,32 +10,34 @@ import org.suai.blamer.issuetracker.IssueTrackerException;
 import org.suai.blamer.output.HTMLPage;
 import org.suai.blamer.output.Screen;
 
+import java.io.FileInputStream;
 import java.io.IOException;
-import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.Properties;
 
 public class Main{
     public static void main(String[] args) throws IOException {
         Options options = new Options();
-        options.addRequiredOption("s", "start", true, "Ticket start number");
-        options.addRequiredOption("e", "end", true, "Ticket end number");
-        options.addRequiredOption("o", "out", true, "You can write either 'screen' or 'html'");
+        options.addOption("s", "start", true, "Ticket start number. Default = 0");
+        options.addOption("e", "end", true, "Ticket end number. Default = 100000");
+        options.addOption("o", "out", true, "You can write either 'screen' or 'html'. Default = screen");
         options.addRequiredOption("l", "login", true, "Login from your github");
         options.addRequiredOption("p", "password", true, "Password from your github");
-        options.addOption("f", "file", true, "The name of the file to display in html");
+        options.addOption("f", "file", true, "The name of the file to display in html. Default = output.html");
+        options.addOption("c", "config properties", true, "Path to file 'config.properties'");
 
         CommandLineParser parser = new DefaultParser();
         CommandLine cmd = null;
         try {
             cmd = parser.parse(options, args);
 
-            int startTicketNum = -1;
-            int endTicketNum = -1;
-            String out = null;
+            int startTicketNum = 0;
+            int endTicketNum = 100000;
+            String out = "screen";
             String htmlFilename = null;
             String login = null;
             String pwd = null;
+            String configPath = "config.properties";
 
             if (cmd.hasOption("s")) {
                 startTicketNum = Integer.parseInt(cmd.getOptionValue("s"));
@@ -53,13 +57,14 @@ public class Main{
             if (cmd.hasOption("p")) {
                 pwd = cmd.getOptionValue("p");
             }
-
-
+            if (cmd.hasOption("c")) {
+                configPath = cmd.getOptionValue("c");
+            }
 
             Properties properties = new Properties();
 
-            InputStream input = Main.class.getClassLoader().getResourceAsStream("config.properties");
-            properties.load(input);
+            FileInputStream fis = new FileInputStream(configPath);
+            properties.load(fis);
 
             String url = properties.getProperty("url");
             String path = properties.getProperty("path");
@@ -99,8 +104,9 @@ public class Main{
         } catch (ParseException pe) {
             HelpFormatter formatter = new HelpFormatter();
             formatter.printHelp("blamer", options);
-            pe.printStackTrace();
         } catch (NullPointerException e){
+            e.printStackTrace();
+        } catch (CheckedIssueException e){
             e.printStackTrace();
         }
     }
