--- src/main/resources/config.properties
@@ -1,5 +1,0 @@
-projectname=guava
-path=/home/mialyshev/guava
-issuetype=github
-url=https://github.com/google/guava
-issuespath =/home/mialyshev/IdeaProjects/graduation_works/scan.txt
